requirements.txt
feedparser==6.0.8
Pillow==9.1.0
pytz==2021.3
requests==2.22.0
